feat: Add support for array properties DML
