+++ packages/core/utils/src/dml/helpers/entity-builder/apply-searchable.ts
@@ -1,0 +1,16 @@
+import { EntityConstructor, PropertyMetadata } from "@medusajs/types"
+import { Searchable } from "../../../dal"
+
+/**
+ * Apply the searchable decorator to the property marked as searchable to enable the free text search
+ */
+export function applySearchable(
+  MikroORMEntity: EntityConstructor<any>,
+  field: PropertyMetadata
+) {
+  if (!field.dataType.options?.searchable) {
+    return
+  }
+
+  Searchable()(MikroORMEntity.prototype, field.fieldName)
+}
